Yet more work on the renderer, small changes to the constant buffer so it remembers its size

+++ amGraphicsDX/Include/amDXConstantBuffer.h
@@ -9,7 +9,7 @@ namespace amEngineSDK {
   {
   public:
     amDXConstantBuffer() = default;
-    amDXConstantBuffer(void* _pCB, const SIZE_T _size);
+    amDXConstantBuffer(void* _pCB, const uint32 _size);
     ~amDXConstantBuffer();
 
     void
